+++ src/app/modules/auth/components/login-network/login-network.component.ts
@@ -1,0 +1,12 @@
+import { AuthService } from 'src/app/modules/auth/auth.service';
+
+import { Component, OnInit } from "@angular/core";
+
+@Component({
+  selector: "app-login-network",
+  templateUrl: "./login-network.component.html",
+})
+export class LoginNetworkComponent implements OnInit {
+  constructor(public authService: AuthService) {}
+  ngOnInit() {}
+}