+++ src/app/shared/constants/index.ts
@@ -1,0 +1,2 @@
+export * from './firebase.constants'
+export * from './global.constants'